backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
supabase>=2.6.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.1
httpx>=0.27.0
